+++ Angular2Emulator/TypeScript/Ng2Emulation/Decorators/Inject.ts
@@ -1,4 +1,4 @@
-﻿import {serviceNormalize} from "Ng2Emulation/Utils/AngularHelpers"
+﻿import {serviceNormalize} from "../Utils/AngularHelpers"
 
 /**
  * Says to angular how and where inject a service or angular 1 component.