+++ Angular2Emulator/TypeScript/Ng2Emulation/Templates/ParserRule.ts
@@ -1,0 +1,3 @@
+﻿export interface ParserRule {
+    processTemplate(template: string): string;
+}