+++ Angular2Emulator/TypeScript/Ng2Emulation/Templates/Rules/NgContentRule.ts
@@ -1,0 +1,7 @@
+﻿import {ParserRule} from "../ParserRule";
+
+export default class NgContentRule implements ParserRule {
+    processTemplate(template: string): string {
+        return template.replace(/ng-content/g, "ng-transclude");
+    }
+}